Dockerizing the app: refactoring the configuration to take optional environment settings.
This commit changes the way the app is configured with a small change: instead of using
json file, it is converted to a module that produces a json. The reason is to maintain
a single configuration file with all the alternatives for default and environment. This way,
the config.js loads the settings as a module and uses it as the default properties.

One way to test is to set the env vars and execute the server:

REDIS_PORT_6379_TCP_ADDR=marcello.redis.localhost REDIS_PORT_6379_TCP_PORT=2345 ./bin/server

*	modified:   config.js
- Loading the new "settings" that was the renamed config.json file.
- Loading it as the "defaults" properties of nconf.

*	deleted:    config.json
*	new file:   settings.js
- As a result of the rename.
- Using the Docker environment variables when using linked containers:
   REDIS_*: when using --link myredis:redis
   COUCHDB_*: when using --link couchdb:couchdb
   PRIVATE_NPM_*: when using --link mynpm:private_npm

*	modified:   Dockerfile
- Simplifying the config and removing the parameter.